Recommit "[CodeView] Add option to disable inline line tables."
This reverts commit 009919aae24e5a18bda4dbaee2c099a1186f217b.
Original commit hash 20880dd4fca90558fecb255de4e1e8d99dc89607

Summary:
This adds a clang option to disable inline line tables. When it is used,
the inliner uses the call site as the location of the inlined function instead of
marking it as an inline location with the function location.

https://reviews.llvm.org/D67723