Remove Support/Options.h, it is unused
It was added in 2014 in 05fe167039bbf0 with one use in Scalarizer.cpp.
That one use was then removed when porting to the new pass manager in
2018 in 02eeef5c2d90b38e401.

While the RFC and the desire to get off of static initializers for
cl::opt all still stand, this code is now dead, and I think we should
delete this code until someone is ready to do the migration.

There were many clients of CommandLine.h that were it transitively
through LLVMContext.h, so I cleaned that up in b3459aab03f4e387c24.

Reviewers: beanz

Differential Revision: https://reviews.llvm.org/D70280